pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "bitsandbytes>=0.45.5",
    "datasets>=3.5.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "langfuse>=2.60.3",
    "llama-index>=0.12.35",
    "llama-index-embeddings-huggingface>=0.5.4",
    "llama-index-retrievers-bm25>=0.5.2",
    "numpy>=2.2.5",
    "openai>=1.76.0",
    "pandas>=2.2.3",
    "peft>=0.15.2",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.0",
    "setuptools>=80.3.1",
    "streamlit>=1.44.1",
    "tensorboard>=2.19.0",
    "tensorboardx>=2.6.2.2",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "trl>=0.17.0",
    "unsloth>=2025.3.3",
    "wheel>=0.45.1",
    "xformers>=0.0.30",
]
